Set cell color by the selected color attribute
This colors cells by a selected attribute, if the attribute is present
in the attribute values collection for the cell's data.

One issue still at hand is that the attribute values are only pulled
down for the selected size attribute, therefore this code only works if
the color attribute is set to the same attribute as the "size by"
attribute.

We'll have to account for multiple attribute selections... shouldn't be
too bad :smile: