Pull down all attribute values from the web API
This allows us access to more than one attribute value at a time. Now we
can color :red_circle: :green_heart: :blue_heart: treemap
:deciduous_tree: cells by attribute values that are different from the
sizing attribute.

It's probably not efficient... but whatevs :wink:

There's no way to pull down just two attribute values within a single
call, so in the future we'll want to use a batch call if we want to
focus down on just the attribute values we want and not the whole set.
Ref: https://pisquare.osisoft.com/thread/12023