Get attributes (with types) before building graphic
The treemap graphic requires the attribute types in order to determine
which color scheme to use: a discrete or sequential color scheme.

These attribute types are not in the `efDataHolder` by the time the
`treemap` function is executing because the attribute types Web API
request is fulfilled via callback and the treemap is built before the
response from that call.

This tries to get the treemap building after the Web API call has
returned.

The attributes and their types should be combined into a batch call
instead of many promises that are all waited upon.